Changed some int data types to NSUInteger.
git-svn-id: https://log4cocoa.svn.sourceforge.net/svnroot/log4cocoa/trunk/source@108 3ad47a68-a840-0410-8984-94a09cb05a02

+++ L4AppenderAttachable.h
@@ -16,7 +16,7 @@
  * @param event the event to be appended.
  * @return the number of appenders the event was appended to.
  */
-- (int) appendLoopOnAppenders: (L4LoggingEvent *) event;
+- (NSUInteger) appendLoopOnAppenders:(L4LoggingEvent *) event;
 
 @end
 // For copyright & license, see COPYRIGHT.txt.
